adk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bat-adk"
version = "2025.12rc1"
authors = [
  { name="Andrea LEONE", email="andrea.leone@bubbleran.com" },
]
description = "Software Development Kit for building AI Agents in BubbleRAN MX-PDK and MX-AI"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "a2a-sdk>=0.3.20",
  "a2a-sdk[http-server]",
  "httpx>=0.28.1",
  "langchain>=0.3.24",
  "langchain-mcp-adapters>=0.1.13,<0.2.0",
  "langchain-nvidia-ai-endpoints>=0.3.9",
  "langchain-openai>=0.3.14",
  "langchain-ollama>=0.3.3",
  "langgraph>=1.0.0",
  "mcp[cli]>=1.17.0",
  "pydantic>=2.10.6",
  "pyyaml>=6.0.1",
  "uvicorn>=0.37.0",
  "bat>=0.3.9",
]
[dependency-groups]
dev = [
  "build",
  "twine",
]
[project.urls]
"Homepage" = "https://bubbleran.com/"
"Repository" = "https://github.com/bubbleran/bat"
